requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
tqdm>=4.65.0 